agents/requirements.txt
openai-agents
python-dotenv
rich